Rework set_init and set_dest, leak proof
This patch reworks and renames the functions set_init to set_create
and set_dest to set_destroy, respectively. In the name of convention,
we also rename the other set functions to be more consistent. See the
doc comments for behavioral changes. In addition, leakproof the API
functions and the test source. A rework of the tests will be coming
soon, which should expose any lingering instability of the API. The
'bugs' file was also added, to keep track of things that need fixing.