Rework test_remove add logging and more testing
This patch reworks the test_remove() function to fit the style of
the new testing. Also, add functionality for logging of errors.
This functionality is enabled in the `debug' rule by default. See the
doc comments on disabling this functionality. Also, add test_create()
and test_destroy(), which test the set_create() and set_destroy()
functions, respectively.