Rework of test_insert() and fixes in set.c
This patch reworks the test source in test_insert() and test_remove().
Test coverage was increased in both of these, and probed for memory
leaks. Testing revealed some vulnerabilities in set_insert() and
set_remove() due to failing to free memory and parameter sanitization,
and this patch fixes those. At this time, the API is completely memleak
free.