Add `copy' parameter
This patch adds the `copy' parameter to the set structure. This is
a pointer to a user-defined function which will copy the user-data
passed as its argument and return a pointer to it. This function
should return NULL if there was an error in copying the data. If this
is the case, the set API knows to return immediately with
corresponding error code. This parameter can be NULL in the set,
however set functions which create a new set using previous set
parameters (set_union, set_intersection, set_difference) will return
immediately.